refactor: confirm redesign footer extract useIsDangerButton
