+++ ui/pages/confirmations/components/confirm/footer/footer.tsx
@@ -1,5 +1,5 @@
 import { ethErrors, serializeError } from 'eth-rpc-errors';
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import {
   Button,
@@ -13,32 +13,22 @@ import { useMMIConfirmations } from '../../../../../hooks/useMMIConfirmations';
 ///: END:ONLY_INCLUDE_IF
 
 import { doesAddressRequireLedgerHidConnection } from '../../../../../selectors';
-import { BlockaidResultType } from '../../../../../../shared/constants/security-provider';
 import {
   rejectPendingApproval,
   resolvePendingApproval,
 } from '../../../../../store/actions';
-import useSignatureSecurityAlertResponse from '../../../hooks/useSignatureSecurityAlertResponse';
 import { confirmSelector } from '../../../selectors';
-import { SecurityAlertResponse } from '../../../types/confirm';
 import { getConfirmationSender } from '../utils';
+import useIsDangerButton from './useIsDangerButton';
 
 const Footer = () => {
   const dispatch = useDispatch();
   const t = useI18nContext();
   const confirm = useSelector(confirmSelector);
-  const [isDangerButton, setIsDangerButton] = useState(false);
+  const isDangerButton = useIsDangerButton();
 
   const { currentConfirmation, isScrollToBottomNeeded } = confirm;
   const { from } = getConfirmationSender(currentConfirmation);
-
-  const currentSecurityAlertId = (
-    currentConfirmation?.securityAlertResponse as SecurityAlertResponse
-  )?.securityAlertId;
-
-  const signatureSecurityAlertResponse = useSignatureSecurityAlertResponse(
-    currentSecurityAlertId,
-  );
 
   ///: BEGIN:ONLY_INCLUDE_IF(build-mmi)
   const { mmiOnSignCallback, mmiSubmitDisabled } = useMMIConfirmations();
@@ -74,13 +64,6 @@ const Footer = () => {
     mmiOnSignCallback();
     ///: END:ONLY_INCLUDE_IF
   }, [currentConfirmation]);
-
-  useEffect(() => {
-    setIsDangerButton(
-      signatureSecurityAlertResponse?.result_type ===
-        BlockaidResultType.Malicious,
-    );
-  }, [signatureSecurityAlertResponse?.result_type]);
 
   return (
     <PageFooter className="confirm-footer_page-footer">
